src/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
idna==3.8
pillow==10.4.0
razorpay==1.4.2
requests==2.32.3
setuptools==74.1.2
sqlparse==0.5.1
stripe==10.10.0
typing_extensions==4.12.2
urllib3==2.2.2
django-cors-headers
drf-yasg
django-quill-editor
PyJWT
django-allauth
cryptography
django-admin-volt
hashids
ua_parser